Create handlers based on the provided scenario
We don't need a separate list of endpoints anymore. This means that no mock endpoints will be created when there are no scenarios defined for those endpoints.

You also don't need to provide method/path anymore when setting a scenario because the method and path can be derived from the scenario.

We still need a nested map on method+path for the active handlers, so that we can override handlers based on method/endpoint and still have multiple handlers active.